Moved files touched by the previous filter bug
