+++ src/components/nvi/table/utils/nvi-curator-aggregations-helpers.ts
