+++ src/hooks/nvi/useFilterInstitutionReportsByUrl.ts
@@ -1,5 +1,5 @@
 import { useSearchParams } from 'react-router';
-import { getNviInstitutionName } from '../../components/nvi/table/nvi-admin-table-helpers';
+import { getNviInstitutionName } from '../../components/nvi/table/utils/nvi-admin-aggregations-helpers';
 import { InstitutionReport } from '../../types/nvi.types';
 
 export const useFilterInstitutionReportsByUrl = (reports: InstitutionReport[]) => {
